crudapp/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-filter==2.3.0
djangorestframework==3.11.0
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
mysqlclient==1.4.6
oauthlib==3.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wincertstore==0.2
zipp==3.1.0 
